Fix padding around notice blocks
Because blocks are using paragraph tags, top and bottom block padding is
dependant on the margin of the global `<p>` style configuration. For
example, in learn2-git theme global paragraph configuration has a
top/bottom margin of 1.7rem (~30px). Current CSS matches this, it
becomes an issue when one wants to use markdown blocks in other themes,
where different font is used with, let's say, Arial with 1.2rem for
paragraph top/bottom margins.

We solve this issue by reseting margin of paragraph tags inside notices
to zero and purely using just padding to provide a spacing.